Verbatim content from BigMistacke (Ware missing)
This was taken directly from @BigMistacke's [post in
Discord](https://discord.com/channels/832448203843567656/1237406903219261500/1237406905622855722).
@BigMistacke hasn't responded to contacts and they've left Bubz's
Discord server, but the content itself is all cc-by-nc-sa licensed per
Posthuman Studios, so this is free to share.

Note that the Ware section was missing in BigMistacke's data (it had
duplicates of the Trait data instead). I'll enter that myself later.